File dialog
For now file dialogs work good enough. Although there is a rare crash that happens with the same image every now and then in the preview on linux. The segmentation fault is not caused by stb image or my memory allocator so either it's the file dialog messing up some addresses or something happens when the textures are loaded and deleted in my vulkan renderer.